+++ scrapli_asyncssh/__init__.py
@@ -1,3 +1,3 @@
 """scrapli asyncssh transport plugin"""
 
-__version__ = "2020.10.10"
+__version__ = "2020.12.23"